+++ src/components/calculator.jsx
@@ -1,25 +1,7 @@
 import { useState } from 'react';
 import calculate from '../logic/calculate';
-import Buttons from './buttons';
 import './calculator.scss';
-
-// Function to create each button of the calculator
-const createButtons = (changeResult) => {
-  const buttonsInput = ['AC', '+/-', '%', '÷', '7', '8', '9', 'x', '4', '5', '6', '-', '1', '2', '3', '+', '0', '.', '='];
-  const buttons = [];
-  buttonsInput.forEach((value) => {
-    buttons.push(
-      <Buttons
-        key={value}
-        value={value}
-        setResult={changeResult}
-        calculate={calculate}
-      />,
-    );
-  });
-
-  return buttons;
-};
+import createButtons from './createButtons';
 
 // Calculator component
 const Calculator = () => {
